reorganizing src folder
+ adding makefile specification (compiler and libraries)

+ changing whole linked list system for faster deletion system

+ so editing htab which use linked list

+ adding font rendering

+ adding openal library (not used yet)

+++ src/event/event.c
@@ -18,5 +18,5 @@ void	initEvent(t_window *window)
 	glfwSetCursorPosCallback(window->ptr, cursorMoveCallback);
 	glfwSetMouseButtonCallback(window->ptr, mouseButtonCallback);
 	glfwSetInputMode(window->ptr, GLFW_CURSOR, GLFW_CURSOR_DISABLED);
-	glfwSetCursorPos(window->ptr, WIDTH / 2, HEIGHT / 2);
+	glfwSetCursorPos(window->ptr, (int)(WIDTH / 2), (int)(HEIGHT / 2));
 }
